pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel-python-client"
version = "2.0.0"
description = "Official Python client for the Sentinel v2 license validation and management API"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [{ name = "Demeng" }]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "PyNaCl>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "respx>=0.21",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "pyright>=1.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src/sentinel"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 99

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "standard"
